fix: paralelizar fetch de API backend y corregir warning de signedness en C++
- backend/server.py: usar ThreadPoolExecutor para consultas en paralelo
- src/main_temp_arg_rpi.cpp: fix -Wsign-compare con static_cast<size_t>

+++ src/main_temp_arg_rpi.cpp
@@ -130,7 +130,7 @@ public:
     void mostrarClimaProvinciaEspecifica(int opcion) {
         inicializarCiudades();
 
-        if (opcion < 1 || opcion > ciudades.size()) {
+        if (opcion < 1 || static_cast<size_t>(opcion) > ciudades.size()) {
             std::cout << "Opción inválida" << std::endl;
             return;
         }
